feat: Split autoupdate_android.sh (2 files) to support new x86_64 arch
This commit renames the files `autoupdate_android.sh` to `autoupdate_android_arm64-v8a.sh` and a new file `autoupdate_android_x86_64.sh` which is almost identical to the former one. I could do this in 1 script, but for a separation of concern I have decided to go this way to make things easier for me to debug / fix.

In my view, all of these build scripts should be replaced with 1 script that can be modified with args to satisfy requirements depending on your preferences or system compability. For example, I might want that script to accept this argument "--build-target android-arm64-v8a", or combine multiple targets. No args would mean it should build for all platforms, but you can imagine that this can get complex or confusing for many because of cross-compilation limitations depending on the host architecture. Hence, I keep it this way for the time being.